requirements.txt
pydantic>=2.0.0
pydantic_ai>=0.1.1
logfire>=0.1.0
streamlit>=1.27.0
atlassian-python-api>=3.41.0
python-dotenv>=1.0.0 